backend/requirements.txt
fastapi>=0.116.1
uvicorn>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.13.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
email-validator>=2.2.0
faker>=30.3.0
logfire==2.11.0
